+++ db/migrate/0016_create_song_contributions.rb
@@ -3,8 +3,8 @@ class CreateSongContributions < ActiveRecord::Migration[7.0]
     create_table :song_contributions do |t|
       t.string :parts     # "vocals, guitar" etc
 
-      t.references :song, null: false, index: true
-      t.references :artist, null: false, index: true
+      t.references :song, null: false, index: true, foreign_key: { on_update: :cascade, on_delete: :cascade }
+      t.references :artist, null: false, index: true, foreign_key: { on_update: :cascade, on_delete: :cascade }
 
       t.timestamps
     end